requirements.txt
absl-py==1.3.0
albumentations==1.3.0
glog==0.3.1
h5py==3.8.0
imagecodecs==2023.3.16
numba==0.56.4
numpy==1.23.4
onnxruntime==1.13.1
opencv_python==4.5.1.48
openpyxl==3.0.9
pandas==1.3.4
PyYAML==6.0
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.2
slideio==2.0.5
tifffile==2022.10.10
torch==2.0.0
torchvision==0.15.1
tqdm==4.64.1